requirements.txt
langgraph>=0.2
langchain>=0.3
langchain-groq>=0.2
langchain-core>=0.3

tweepy>=4.14
praw>=7.7
tavily-python>=0.5

transformers>=4.40
torch>=2.2

streamlit>=1.32
pandas>=2.2
plotly>=5.20
python-dotenv>=1.0
schedule>=1.2